refactor: useActiveSection 훅 분리
