+++ src/components/DetailSection.tsx
@@ -1,9 +1,9 @@
 'use client';
 
-import { useState, useEffect } from 'react';
 import TextBlock from '@/components/TextBlock';
 import MediaBlock from '@/components/MediaBlock';
 import Image from 'next/image';
+import useActiveSection from '@/hooks/useActiveSection';
 
 interface Contents {
   contents: {
@@ -15,39 +15,7 @@ interface Contents {
 }
 
 const DetailSection = ({ contents }: Contents) => {
-  const [activeSection, setActiveSection] = useState<number | null>(null);
-
-  useEffect(() => {
-    const handleScroll = () => {
-      const sections = document.querySelectorAll('[data-order]');
-      const viewportMiddle = window.scrollY + window.innerHeight / 2;
-
-      let isActiveSection = false;
-
-      sections.forEach((section) => {
-        const rect = section.getBoundingClientRect();
-        const sectionTop = window.scrollY + rect.top;
-        const sectionBottom = sectionTop + rect.height;
-
-        if (viewportMiddle >= sectionTop && viewportMiddle <= sectionBottom) {
-          const index = parseInt(section.getAttribute('data-order') || '0', 10);
-          setActiveSection(index);
-          isActiveSection = true;
-        }
-      });
-
-      if (!isActiveSection) {
-        setActiveSection(null);
-      }
-    };
-
-    window.addEventListener('scroll', handleScroll);
-    handleScroll();
-
-    return () => {
-      window.removeEventListener('scroll', handleScroll);
-    };
-  }, []);
+  const activeSection = useActiveSection();
 
   return (
     <div className="flex w-full items-center justify-center">
